pyproject.toml
[project]
name = "cirno"
version = "0.1.0"
description = "cirno"
readme = "README.md"
requires-python = ">=3.8, <4.0"

[tool.nonebot]
adapters = [
    { name = "mirai2", module_name = "nonebot.adapters.mirai2" }
]
plugins = []
plugin_dirs = ["plugins"]
builtin_plugins = ["echo"]

[tool.poetry]
name = "cirno"
version = "0.1.0"
description = ""
authors = ["KosWu <ws00298046@163.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
nonebot-adapter-mirai2 = "^0.0.22"
nonebot-adapter-console = "^0.4.0"
nonebot2 = {extras = ["websockets"], version = "^2.1.1"}
typer = {extras = ["all"], version = "^0.9.0"}
pillow = "^10.0.1"
pydub = "^0.25.1"


[tool.poetry.group.dev.dependencies]
black = {version = "^23.9.1", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
